Validations

+++ application/controllers/Home.php
@@ -56,15 +56,20 @@ class Home extends CI_Controller {
     }
 
     public function accident_type_register() {
-      $id = $this->input->post('id');
-      $name = $this->input->post('name');
-
+      $this->load->library('form_validation');
       $this->load->model('accident_type');
 
-      if (strlen($id) > 0) {
-        $this->accident_type->update($id, $name);
-      } else {
-        $this->accident_type->create($name);
+      $this->form_validation->set_rules('name', 'Email', 'required');
+
+      if ($this->form_validation->run()) {
+        $id = $this->input->post('id');
+        $name = $this->input->post('name');
+
+        if (strlen($id) > 0) {
+          $this->accident_type->update($id, $name);
+        } else {
+          $this->accident_type->create($name);
+        }
       }
 
       redirect(base_url('home/accident_types'), 'location');
@@ -72,10 +77,15 @@ class Home extends CI_Controller {
     }
 
     public function accident_type_delete() {
-      $id = $this->input->post('id');
+      $this->load->library('form_validation');
+      $this->form_validation->set_rules('id', 'Clave', 'required');
 
-      $this->load->model('accident_type');
-      $this->accident_type->delete($id);
+      if ($this->form_validation->run()) {
+        $id = $this->input->post('id');
+
+        $this->load->model('accident_type');
+        $this->accident_type->delete($id);
+      }
 
       redirect(base_url('home/accident_types'), 'location');
       return;
@@ -144,18 +154,27 @@ class Home extends CI_Controller {
     }
 
     public function user_register() {
-      $id = $this->input->post('id');
-      $name = $this->input->post('name');
-      $password = $this->input->post('password');
-      $email = $this->input->post('email');
-      $is_admin = $this->input->post('is_admin');
+      $this->load->library('form_validation');
 
-      $this->load->model('user');
+      $this->form_validation->set_rules('name', 'Nombre', 'required');
+      $this->form_validation->set_rules('password', 'Contraseña', 'required');
+      $this->form_validation->set_rules('email', 'Email', 'required');
+      $this->form_validation->set_rules('is_admin', 'is_admin', 'required');
 
-      if (strlen($id) > 0) {
-        $this->user->update($id, $name, $email, $password, $is_admin);
-      } else {
-        $this->user->create($name, $email, $password, $is_admin);
+      if ($this->form_validation->run()) {
+        $id = $this->input->post('id');
+        $name = $this->input->post('name');
+        $password = $this->input->post('password');
+        $email = $this->input->post('email');
+        $is_admin = $this->input->post('is_admin');
+
+        $this->load->model('user');
+
+        if (strlen($id) > 0) {
+          $this->user->update($id, $name, $email, $password, $is_admin);
+        } else {
+          $this->user->create($name, $email, $password, $is_admin);
+        }
       }
 
       redirect(base_url('home/users'), 'location');
@@ -163,10 +182,15 @@ class Home extends CI_Controller {
     }
 
     public function user_delete() {
-      $id = $this->input->post('id');
+      $this->load->library('form_validation');
+      $this->form_validation->set_rules('id', 'Clave', 'required');
 
-      $this->load->model('user');
-      $this->user->delete($id);
+      if ($this->form_validation->run()) {
+        $id = $this->input->post('id');
+
+        $this->load->model('user');
+        $this->user->delete($id);
+      }
 
       redirect(base_url('home/users'), 'location');
       return;
